playright_v1/requirements.txt
playwright==1.42.0
pytest-playwright==0.4.4
Pillow>=10.0.0
imagehash>=4.3.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
requests>=2.31.0
mss>=9.0.0
urllib3>=2.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
